+++ CHANGELOG.md
@@ -5,10 +5,17 @@
 ### Added
 
 - Tailwind
+- Buttons
+- Inputs: text input, textarea, select, progressbar, checkbox, radio
+- Panels and cards
+- Customized cursors
 
 ### Changed
 
+- Color palette
+- Playground to demo
 - Refactored all existing CSS rules
+- Size limit: 35 kb -> 25 kb
 
 ## [0.3.0] - 2026-03-17